New version with several improvements and bugfixes

+++ codenerix/lib/pylock.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 #
 # django-codenerix
 #
@@ -18,32 +17,35 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-'''
+"""
 Library to handle lockers over files
-'''
+"""
 
 __version__ = "2015061200"
 
-import os
 import fcntl
-import unittest
+import os
 import tempfile
+import unittest
 
-__all__ = [ "pylock" , "AlreadyLocked" ]
+__all__ = ["pylock", "PyLock", "AlreadyLocked"]
 
-class pylock:
-    '''
+
+class PyLock:
+    """
     Function to control locking flags over a file
-    '''
+    """
 
     def __init__(self, lockfile, locktype):
-        '''
+        """
         Parameters:
         - `lockfile`: name of the file to check/apply the locking.
         - `locktype`: possible values are:
-                wait: on a call to lock() function, the system will wait to get the locker
-                lock: on a call to lock() function, if locked the system will raise an AlreadyLocked exception
-        '''
+                wait: on a call to lock() function, the system will wait to
+                        get the locker
+                lock: on a call to lock() function, if locked the system will
+                        raise an AlreadyLocked exception
+        """
 
         # Save config
         self.__lockfile = lockfile
@@ -52,27 +54,29 @@ class pylock:
 
         # Check file exists and create it if it does not
         if not os.path.exists(lockfile):
-            file = open(lockfile, 'w')
+            file = open(lockfile, "w")
             file.close()
 
         # Check locktype
-        if locktype not in ['wait','lock']:
+        if locktype not in ["wait", "lock"]:
             raise TypeError("Locking type unkown")
 
     def __del__(self):
-        '''
+        """
         when dying make sure the lock become free
-        '''
+        """
         # If file was open, close it and delete it!
         if self.__fd:
             self.__fd.close()
 
     def lock(self):
-        '''
+        """
         Try to get locked the file
-        - the function will wait until the file is unlocked if 'wait' was defined as locktype
-        - the funciton will raise AlreadyLocked exception if 'lock' was defined as locktype
-        '''
+        - the function will wait until the file is unlocked if 'wait' was
+            defined as locktype
+        - the funciton will raise AlreadyLocked exception if 'lock' was
+            defined as locktype
+        """
 
         # Open file
         self.__fd = open(self.__lockfile, "w")
@@ -84,14 +88,14 @@ class pylock:
         elif self.__locktype == "lock":
             # Try to get the locker if can not raise an exception
             try:
-                fcntl.flock(self.__fd.fileno(), fcntl.LOCK_EX|fcntl.LOCK_NB)
-            except IOError:
+                fcntl.flock(self.__fd.fileno(), fcntl.LOCK_EX | fcntl.LOCK_NB)
+            except OSError:
                 raise AlreadyLocked("File is already locked")
 
     def free(self):
-        '''
+        """
         Set the locked file free
-        '''
+        """
 
         # Close file
         self.__fd.close()
@@ -99,25 +103,28 @@ class pylock:
         # Delete it
         try:
             os.unlink(self.__lockfile)
-        except:
+        except Exception:
             pass
 
+
+# Compatibility with older versions
+pylock = PyLock
 
 
 # Exceptions classes
 class AlreadyLocked(Exception):
-
-    def __init__(self,string):
+    def __init__(self, string):
         self.string = string
 
     def __str__(self):
         return self.string
 
+
 # Testing
-class test_pylock(unittest.TestCase):
-    '''
+class TestPylock(unittest.TestCase):
+    """
     Testing class for pylock
-    '''
+    """
 
     def setUp(self):
         pass
@@ -127,10 +134,10 @@ class test_pylock(unittest.TestCase):
         f = tempfile.NamedTemporaryFile(delete=False)
         f.close()
 
-        locker1 = pylock(f.name, 'lock')
+        locker1 = PyLock(f.name, "lock")
         locker1.lock()
 
-        locker2 = pylock(f.name, 'lock')
+        locker2 = PyLock(f.name, "lock")
         self.assertRaises(AlreadyLocked, locker2.lock)
         locker1.free()
         locker2.lock()
@@ -138,9 +145,11 @@ class test_pylock(unittest.TestCase):
         locker2.free()
 
         # Remove the temporal file
-        os.unlink(f.name)
+        if os.path.exists(f.name):
+            os.unlink(f.name)
+            raise OSError("File should be already deleted by the library")
 
 
 # Base call
-if __name__ == '__main__':
+if __name__ == "__main__":
     unittest.main()
